feat.cmd,config,accrual,service.worker,psql,db: accrual HTTP client + deps-injected Service + worker DB ops + int balances
Context:
- Нужно связать воркер с реальным HTTP-клиентом accrual-сервиса и сделать сервис “собираемым” через явные зависимости (deps), чтобы не раздувать сигнатуру New и не терять контроль над обязательными компонентами.
- Для корректной работы воркера требуется репозиторий, который умеет:
  - выбирать pending-заказы по next_sync_at,
  - обновлять next_sync_at отдельно от статуса,
  - атомарно помечать заказ PROCESSED и начислять баланс пользователю в транзакции.
- Для упрощения доменной модели и sqlc-маппинга перевести денежные поля на INTEGER, а accrual в ответе сервиса сделать nullable.

Changes:
- Makefile:
  - run:
    - Порт локального accrual-бинарника изменён на localhost:8081.

- cmd/gophermart:
  - main.go:
    - Подключен accrualclient.New(cfg.AccrualServiceAddress).
    - gophermart.New переведён на deps-подход:
      - New(cfg, ServiceDeps{Hasher, UserCRUD, WorkerDB, AddOrdered, AccrualClient}).
    - Обработка ошибок создания сервиса: теперь New возвращает (*Service, error).

- internal/config:
  - config.go:
    - Добавлен параметр AccrualServiceAddress:
      - дефолт "http://localhost:8081/"
      - флаг `-r` (accrual service address)
      - env `ACCRUAL_SYSTEM_ADDRESS`.

- internal/accrual_client:
  - client.go:
    - Добавлен HTTP-клиент accrual-сервиса:
      - GET {baseUrl}/api/orders/{number}
      - 429 → gophermart.ErrToManyRequests
      - 204 No Content → трактуется как NEW (REGISTERED) для текущего заказа
      - JSON распарсивается в gophermart.AccrualResponse
      - Маппинг статусов accrual → внутренние: REGISTERED→NEW, INVALID/PROCESSING/PROCESSED.

- internal/service/gophermart:
  - interfaces.go:
    - AccrualResponse.Accrual изменён на *float64 (nullable, чтобы корректно обработать отсутствие начисления).
    - ListUpdateApplyAccrual:
      - ListPending теперь принимает limit int32 (согласовано с sqlc/pgx).
      - Добавлен метод UpdateSyncTime(number, nextSync) для обновления next_sync_at без изменения статуса.
  - service.go:
    - Service расширен полями worker/workerDb/clientAccrual.
    - Введён ServiceDeps:
      - Hasher, UserCRUD, AddOrdered, WorkerDB, AccrualClient.
    - New(cfg, deps) теперь:
      - валидирует обязательные зависимости и возвращает error с причиной,
      - создаёт worker через newWorker(deps.AccrualClient, deps.WorkerDB).
    - Добавлены методы жизненного цикла:
      - Start() → worker.Run()
      - Stop() → worker.Stop()

  - worker.go:
    - Улучшено логирование запуска батча:
      - при наличии заказов логирует count_orders и список номеров.
    - PROCESSED:
      - ApplyAccrual теперь использует *responseAccrual.Accrual (dereference).
    - Если статус не изменился:
      - вместо UpdateFromAccrual(status="") теперь вызывается UpdateSyncTime.
    - Добавлен helper listOrdersString для формирования списка номеров.

- internal/repository/psql:
  - for_worker.go:
    - Добавлена реализация интерфейса воркера на репозитории:
      - ListPending(limit, statuses, timeSync) → выборка pending по next_sync_at.
      - UpdateFromAccrual(number, status, nextSync) → обновляет status + next_sync_at.
      - UpdateSyncTime(number, nextSync) → обновляет только next_sync_at.
      - ApplyAccrual(number, accrual, userID):
        - выполняется в транзакции (InTx):
          - MarkOrderProcessed (status=PROCESSED, accrual, next_sync_at=NULL) + RETURNING,
          - AddToUserBalanceUpsert (upsert баланса).
  - intx.go:
    - Добавлен helper InTx(ctx, func(*Repo) error):
      - Begin/Commit/Rollback,
      - прокидывает Queries.WithTx(tx) для sqlc.

  - queries/worker.sql:
    - Добавлены запросы для воркера:
      - ListPending (status IN statuses AND (next_sync_at IS NULL OR next_sync_at <= now)).
      - UpdateFromAccrual.
      - UpdateSyncTime.
      - MarkOrderProcessed (idempotent: status <> PROCESSED, фильтр по user_id).
      - AddToUserBalanceUpsert (accumulate balance).
  - query/worker.sql.go:
    - Добавлен сгенерированный sqlc-код под worker.sql.
  - query/models.go:
    - Приведение типов под новую схему:
      - OrderNumber.Accrual: pgtype.Int4, добавлен NextSyncAt.
      - UserBalance.Balance/Withdrawn: int32.
      - Withdraw.Summa: int32.

- migrations/schema:
  - 00001_init_table_in_database.sql:
    - Денежные поля переведены на INTEGER:
      - order_numbers.accrual: INTEGER
      - user_balance.balance/withdrawn: INTEGER
      - withdraws.summa: INTEGER

Implementation:
- Внешний accrual-интеграционный слой выделен в отдельный пакет internal/accrual_client и подключается в main через cfg.AccrualServiceAddress.
- Сервис переведён на DI через ServiceDeps:
  - проще расширять зависимостями (worker db / accrual client) без роста списка аргументов,
  - early-fail с понятными ошибками при отсутствии обязательных зависимостей.
- Воркера “замкнули” на БД через специализированный интерфейс ListUpdateApplyAccrual:
  - выборка pending по next_sync_at,
  - раздельные операции “обновить статус” vs “сдвинуть next_sync_at”,
  - транзакционное применение accrual + баланс.
- Типы денег/начислений упрощены:
  - в БД целые значения,
  - в ответе accrual accrual может отсутствовать → *float64.

Tests:
- TODO:
  - Unit-тесты accrualclient.GetOrder:
    - 204 → Status=NEW и OrderNumber=number.
    - 429 → ErrToManyRequests.
    - маппинг REGISTERED→NEW и неизвестный статус → error.
  - Unit/интеграционные тесты Repo.ApplyAccrual (InTx):
    - MarkOrderProcessed не обновляет уже PROCESSED.
    - баланс увеличивается на accrual.
  - Тесты Service.New (deps validation) + Start/Stop воркера.
  - Тесты worker.checkAndUpdate:
    - UpdateSyncTime вызывается при неизменном статусе,
    - ApplyAccrual использует accrual только при PROCESSED и accrual != nil.

+++ migrations/schema/00001_init_table_in_database.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS order_numbers (
     "number" VARCHAR(255) NOT NULL,
     user_id INTEGER NOT NULL,
     "status" VARCHAR(30) NOT NULL,
-    accrual NUMERIC(12, 2),
+    accrual INTEGER,
     uploaded_at TIMESTAMPTZ NOT NULL,
     next_sync_at TIMESTAMPTZ,
 
@@ -31,8 +31,8 @@ WHERE "status" IN ('NEW', 'PROCESSING');
 CREATE TABLE IF NOT EXISTS user_balance (
         id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
         user_id INTEGER NOT NULL,
-        balance NUMERIC(12,2) NOT NULL DEFAULT 0,
-        withdrawn NUMERIC(12,2) NOT NULL DEFAULT 0,
+        balance INTEGER NOT NULL DEFAULT 0,
+        withdrawn INTEGER NOT NULL DEFAULT 0,
 
         CHECK (balance >= 0 AND withdrawn >= 0),
         CONSTRAINT user_balance_user_uk UNIQUE (user_id),
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS withdraws (
     id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     user_id INTEGER NOT NULL,
     order_number VARCHAR(255) NOT NULL,
-    summa NUMERIC(12,2) NOT NULL,
+    summa INTEGER NOT NULL,
     processed_at TIMESTAMPTZ NOT NULL,
